Use ip4s for Server address
