+++ blaze-server/src/test/scala/org/http4s/blaze/server/BlazeServerSuite.scala
@@ -105,13 +105,13 @@ class BlazeServerSuite extends Http4sSuite {
 
   def get(server: Server, path: String): IO[String] = IO.blocking {
     Source
-      .fromURL(new URL(s"http://127.0.0.1:${server.address.getPort}$path"))
+      .fromURL(new URL(s"http://${server.address}$path"))
       .getLines()
       .mkString
   }
 
   def getStatus(server: Server, path: String): IO[Status] = {
-    val url = new URL(s"http://127.0.0.1:${server.address.getPort}$path")
+    val url = new URL(s"http://${server.address}$path")
     for {
       conn <- IO.blocking(url.openConnection().asInstanceOf[HttpURLConnection])
       _ = conn.setRequestMethod("GET")
@@ -122,7 +122,7 @@ class BlazeServerSuite extends Http4sSuite {
   }
 
   def post(server: Server, path: String, body: String): IO[String] = IO.blocking {
-    val url = new URL(s"http://127.0.0.1:${server.address.getPort}$path")
+    val url = new URL(s"http://${server.address}$path")
     val conn = url.openConnection().asInstanceOf[HttpURLConnection]
     val bytes = body.getBytes(StandardCharsets.UTF_8)
     conn.setRequestMethod("POST")
@@ -138,7 +138,7 @@ class BlazeServerSuite extends Http4sSuite {
       boundary: String,
       body: String): IO[String] =
     IO.blocking {
-      val url = new URL(s"http://127.0.0.1:${server.address.getPort}$path")
+      val url = new URL(s"http://${server.address}$path")
       val conn = url.openConnection().asInstanceOf[HttpURLConnection]
       val bytes = body.getBytes(StandardCharsets.UTF_8)
       conn.setRequestMethod("POST")
