--- client/jvm/src/test/scala/org/http4s/client/ClientRouteTestBatteryPlatform.scala
@@ -1,38 +1,0 @@
-/*
- * Copyright 2014 http4s.org
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.http4s.client
-
-import cats.effect.IO
-import cats.effect.kernel.Resource
-import org.http4s.Uri
-import org.http4s.server.Server
-
-import java.net.InetSocketAddress
-
-trait ClientRouteTestBatteryPlatform { self: ClientRouteTestBattery =>
-
-  def serverResource: Resource[IO, Server] = Resource.pure(new Server {
-    override val address: InetSocketAddress =
-      InetSocketAddress.createUnresolved("localhost", 8888)
-    override def isSecure: Boolean = false
-  })
-
-  def url(path: String): IO[Uri] =
-    server().map(s =>
-      Uri.fromString(s"http://${s.address.getHostName}:${s.address.getPort}$path").yolo)
-
-}
