+++ client/shared/src/test/scala/org/http4s/client/ClientRouteTestBattery.scala
@@ -19,22 +19,33 @@ package client
 
 import cats.effect._
 import cats.syntax.all._
+import com.comcast.ip4s.Host
+import com.comcast.ip4s.SocketAddress
 import fs2._
-import java.util.Arrays
-import java.util.Locale
 import org.http4s.client.dsl.Http4sClientDsl
 import org.http4s.client.testroutes.GetRoutes
 import org.http4s.dsl.io._
-import org.http4s.multipart.{Multipart, Part}
+import org.http4s.multipart.Multipart
+import org.http4s.multipart.Part
+import org.http4s.server.Server
+
+import java.util.Arrays
+import java.util.Locale
 import scala.concurrent.duration._
 
-abstract class ClientRouteTestBattery(name: String)
-    extends Http4sSuite
-    with Http4sClientDsl[IO]
-    with ClientRouteTestBatteryPlatform {
+abstract class ClientRouteTestBattery(name: String) extends Http4sSuite with Http4sClientDsl[IO] {
   val timeout = 20.seconds
 
   def clientResource: Resource[IO, Client[IO]]
+
+  def serverResource: Resource[IO, Server] = Resource.pure(new Server {
+    override def address: SocketAddress[Host] =
+      SocketAddress.fromStringHostname(s"localhost:8888").get
+    override def isSecure: Boolean = false
+  })
+
+  def url(path: String): IO[Uri] =
+    server().map(s => Uri.fromString(s"http://${s.address.toString}$path").yolo)
 
   val server = resourceSuiteDeferredFixture("server", serverResource)
   val client = resourceSuiteDeferredFixture("client", clientResource)
