+++ server/shared/src/main/scala/org/http4s/server/Server.scala
@@ -17,10 +17,32 @@
 package org.http4s
 package server
 
+import com.comcast.ip4s
 import org.log4s.getLogger
 
-abstract class Server extends ServerPlatform {
+abstract class Server {
   private[server] val logger = getLogger
+
+  def address: ip4s.SocketAddress[ip4s.Host]
+
+  def baseUri: Uri =
+    Uri(
+      scheme = Some(if (isSecure) Uri.Scheme.https else Uri.Scheme.http),
+      authority = Some(
+        Uri.Authority(
+          host = address.host match {
+            case ipv4: ip4s.Ipv4Address =>
+              Uri.Ipv4Address(ipv4)
+            case ipv6: ip4s.Ipv6Address =>
+              Uri.Ipv6Address(ipv6)
+            case weird =>
+              logger.warn(s"Unexpected address type ${weird.getClass}: $weird")
+              Uri.RegName(weird.toString())
+          },
+          port = Some(address.port.value)
+        )),
+      path = Uri.Path.Root
+    )
 
   def isSecure: Boolean
 }
